chore: move types definitions into api
