+++ api/doc/agg-query/schema.json
