+++ api/doc/agg-result/schema.json
