+++ api/doc/index.ts
@@ -1,0 +1,9 @@
+import * as aggQuery from './agg-query/index.js'
+import * as aggResult from './agg-result/index.js'
+import * as listResponse from './list-response/index.js'
+
+export default {
+  aggQuery,
+  aggResult,
+  listResponse
+}
