+++ api/types/daily-api-metric/schema.json
