+++ api/types/index.ts
@@ -1,0 +1,1 @@
+export type { DailyApiMetric } from './daily-api-metric.js'
