+++ shared/index.js
@@ -1,4 +1,4 @@
-export * as aggQueryType from './agg-query/index.js'
-export * as aggResultType from './agg-result/index.js'
-export * as listResponseType from './list-response/index.js'
+export * as aggQueryType from '../api/doc/agg-query/index.js'
+export * as aggResultType from '../api/doc/agg-result/index.js'
+export * as listResponseType from '../api/doc/list-response/index.js'
 export * as dailyApiMetricType from './daily-api-metric/index.js'
